Refactor regridder unit tests to use subclassing.
